cosmochroma-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python>=4.8.1.78
numpy>=1.26.0
scipy>=1.13.0
scikit-image>=0.24.0
pandas>=2.2.0
tensorflow>=2.16.0
mediapipe>=0.10.7
python-dotenv>=1.0.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
